refactor: rename to clouddeploy
